Fix comment formatting

+++ deploy_script/deploy.sh
@@ -281,7 +281,8 @@ then
   wget -q -P $hasp_temp_dir https://github.com/zonko16/Custom-Pages-for-HASwitchplate/raw/dev/packages_3.2in/3.2_packages.tar.gz
   tar -zxf $hasp_temp_dir/3.2_packages.tar.gz -C $hasp_temp_dir
   rm $hasp_temp_dir/3.2_packages.tar.gz
-else
+elif [[ "$panel_size"  == "2.4" ]]
+then
   wget -q -P $hasp_temp_dir https://github.com/zonko16/Custom-Pages-for-HASwitchplate/raw/dev/packages_2.4in/2.4_packages.tar.gz
   tar -zxf $hasp_temp_dir/2.4_packages.tar.gz -C $hasp_temp_dir
   rm $hasp_temp_dir/2.4_packages.tar.gz
